+++ core/java/src/net/i2p/crypto/eddsa/math/ed25519/Ed25519LittleEndianEncoding.java
@@ -197,7 +197,7 @@ public class Ed25519LittleEndianEncoding extends Encoding {
         result |= (in[offset++] & 0xff) << 8;
         result |= (in[offset++] & 0xff) << 16;
         result |= in[offset] << 24;
-        return ((long) result) & 0xffffffffL;
+        return ( result) & 0xffffffffL;
     }
 
     /**
@@ -230,35 +230,35 @@ public class Ed25519LittleEndianEncoding extends Encoding {
         long carry9;
 
         // Remember: 2^255 congruent 19 modulo p
-        carry9 = (h9 + (long) (1 << 24)) >> 25;
+        carry9 = (h9 +  (1 << 24)) >> 25;
         h0 += carry9 * 19;
         h9 -= carry9 << 25;
-        carry1 = (h1 + (long) (1 << 24)) >> 25;
+        carry1 = (h1 +  (1 << 24)) >> 25;
         h2 += carry1;
         h1 -= carry1 << 25;
-        carry3 = (h3 + (long) (1 << 24)) >> 25;
+        carry3 = (h3 +  (1 << 24)) >> 25;
         h4 += carry3;
         h3 -= carry3 << 25;
-        carry5 = (h5 + (long) (1 << 24)) >> 25;
+        carry5 = (h5 +  (1 << 24)) >> 25;
         h6 += carry5;
         h5 -= carry5 << 25;
-        carry7 = (h7 + (long) (1 << 24)) >> 25;
+        carry7 = (h7 +  (1 << 24)) >> 25;
         h8 += carry7;
         h7 -= carry7 << 25;
 
-        carry0 = (h0 + (long) (1 << 25)) >> 26;
+        carry0 = (h0 +  (1 << 25)) >> 26;
         h1 += carry0;
         h0 -= carry0 << 26;
-        carry2 = (h2 + (long) (1 << 25)) >> 26;
+        carry2 = (h2 +  (1 << 25)) >> 26;
         h3 += carry2;
         h2 -= carry2 << 26;
-        carry4 = (h4 + (long) (1 << 25)) >> 26;
+        carry4 = (h4 +  (1 << 25)) >> 26;
         h5 += carry4;
         h4 -= carry4 << 26;
-        carry6 = (h6 + (long) (1 << 25)) >> 26;
+        carry6 = (h6 +  (1 << 25)) >> 26;
         h7 += carry6;
         h6 -= carry6 << 26;
-        carry8 = (h8 + (long) (1 << 25)) >> 26;
+        carry8 = (h8 +  (1 << 25)) >> 26;
         h9 += carry8;
         h8 -= carry8 << 26;
